+++ src/BaseSection.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Towa\Acf;
+namespace DigitOne\Acf;
 
 /**
  * @deprecated (use Fields\Layout instead)