+++ src/Fields/Number.php
@@ -1,7 +1,7 @@
 <?php
-namespace Towa\Acf\Fields;
+namespace DigitOne\Acf\Fields;
 
-use Towa\Acf\BaseField;
+use DigitOne\Acf\BaseField;
 
 class Number extends BaseField
 {